refactor: extract duplicated method to a trait
