+++ test/unit/models/classes/LtiRestApiServiceTest.php
@@ -27,6 +27,7 @@ use oat\generis\model\data\Ontology;
 use oat\generis\test\ServiceManagerMockTrait;
 use oat\taoLti\models\classes\ConsumerService;
 use oat\taoLti\models\classes\LtiRestApiService;
+use oat\taoLti\test\unit\OntologyMockTrait;
 use PHPUnit\Framework\TestCase;
 use ReflectionException;
 use ReflectionMethod;
@@ -35,6 +36,7 @@ use tao_models_classes_CrudService;
 class LtiRestApiServiceTest extends TestCase
 {
     use ServiceManagerMockTrait;
+    use OntologyMockTrait;
 
     /**
      * @throws ReflectionException
@@ -59,23 +61,5 @@ class LtiRestApiServiceTest extends TestCase
 
         $this->assertInstanceOf(core_kernel_classes_Class::class, $rootClass);
         $this->assertEquals(ConsumerService::CLASS_URI, $rootClass->getUri());
-    }
-
-    /**
-     * @return core_kernel_persistence_smoothsql_SmoothModel
-     */
-    protected function getOntologyMock(): core_kernel_persistence_smoothsql_SmoothModel
-    {
-        $model = new core_kernel_persistence_smoothsql_SmoothModel([
-            core_kernel_persistence_smoothsql_SmoothModel::OPTION_PERSISTENCE => 'mockSql',
-            core_kernel_persistence_smoothsql_SmoothModel::OPTION_READABLE_MODELS => [2,3],
-            core_kernel_persistence_smoothsql_SmoothModel::OPTION_WRITEABLE_MODELS => [2],
-            core_kernel_persistence_smoothsql_SmoothModel::OPTION_NEW_TRIPLE_MODEL => 2,
-        ]);
-        $this->getServiceManagerMock([
-            Ontology::SERVICE_ID => $model,
-        ]);
-
-        return $model;
     }
 }
