+++ test/unit/models/classes/ProviderServiceTest.php
@@ -22,16 +22,15 @@
 namespace oat\taoLti\test\unit\models\classes;
 
 use core_kernel_classes_Class;
-use core_kernel_persistence_smoothsql_SmoothModel;
-use oat\generis\model\data\Ontology;
 use oat\generis\test\ServiceManagerMockTrait;
-use oat\generis\test\unit\OntologyMockTest;
 use oat\taoLti\models\classes\ProviderService;
+use oat\taoLti\test\unit\OntologyMockTrait;
 use PHPUnit\Framework\TestCase;
 
 class ProviderServiceTest extends TestCase
 {
     use ServiceManagerMockTrait;
+    use OntologyMockTrait;
 
     public function testGetRootClass()
     {
@@ -42,23 +41,5 @@ class ProviderServiceTest extends TestCase
 
         $this->assertInstanceOf(core_kernel_classes_Class::class, $rootClass);
         $this->assertEquals(ProviderService::CLASS_URI, $rootClass->getUri());
-    }
-
-    /**
-     * @return core_kernel_persistence_smoothsql_SmoothModel
-     */
-    protected function getOntologyMock(): core_kernel_persistence_smoothsql_SmoothModel
-    {
-        $model = new core_kernel_persistence_smoothsql_SmoothModel([
-            core_kernel_persistence_smoothsql_SmoothModel::OPTION_PERSISTENCE => 'mockSql',
-            core_kernel_persistence_smoothsql_SmoothModel::OPTION_READABLE_MODELS => [2,3],
-            core_kernel_persistence_smoothsql_SmoothModel::OPTION_WRITEABLE_MODELS => [2],
-            core_kernel_persistence_smoothsql_SmoothModel::OPTION_NEW_TRIPLE_MODEL => 2,
-        ]);
-        $this->getServiceManagerMock([
-            Ontology::SERVICE_ID => $model,
-        ]);
-
-        return $model;
     }
 }
